♻️ refactor: LendingBadge
