+++ pages/de-fi.tsx
@@ -1,7 +1,7 @@
 import { NextPage } from 'next';
 
 import { SEO } from '@/components';
-import DeFi from '@/views/defi';
+import DeFi from '@/views/de-fi';
 
 const DefiPage: NextPage = () => (
   <>
