+++ views/de-fi/components/all/components/info-box/info-box.types.ts
