+++ views/de-fi/components/all/components/lending-badge/lending-badge.types.ts
@@ -1,4 +1,4 @@
 export interface ILendingBadge {
   platform: string;
-  badgeColor: string;
+  imageUrl: string;
 }
