+++ views/de-fi/components/all/components/lending-card/lending-card.types.ts
@@ -1,6 +1,6 @@
 export interface ILendingCard {
   platform: string;
-  badgeColor: string;
+  imageUrl: string;
   asset: string;
   apr: string;
   tvl: string;
