+++ views/de-fi/defi-content.tsx
